Use cobra/viper for flag parsing
